+++ src/test/java/com/maxmind/minfraud/request/EventTest.java
@@ -3,12 +3,22 @@ package com.maxmind.minfraud.request;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import com.maxmind.minfraud.request.Event.Builder;
+import com.maxmind.minfraud.request.Event.Party;
 import com.maxmind.minfraud.request.Event.Type;
 import java.time.ZonedDateTime;
 import java.util.Date;
 import org.junit.jupiter.api.Test;
 
 public class EventTest {
+
+    @Test
+    public void testParty() {
+        Event event = new Builder().party(Party.AGENT).build();
+        assertEquals(Party.AGENT, event.getParty());
+
+        event = new Builder().party(Party.CUSTOMER).build();
+        assertEquals(Party.CUSTOMER, event.getParty());
+    }
 
     @Test
     public void testTransactionId() {
@@ -43,5 +53,11 @@ public class EventTest {
 
         event = new Builder().type(Type.PAYOUT_CHANGE).build();
         assertEquals(Type.PAYOUT_CHANGE, event.getType());
+
+        event = new Builder().type(Type.CREDIT_APPLICATION).build();
+        assertEquals(Type.CREDIT_APPLICATION, event.getType());
+
+        event = new Builder().type(Type.FUND_TRANSFER).build();
+        assertEquals(Type.FUND_TRANSFER, event.getType());
     }
 }