+++ src/test/java/com/maxmind/minfraud/request/PaymentTest.java
@@ -4,10 +4,23 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.maxmind.minfraud.request.Payment.Builder;
+import com.maxmind.minfraud.request.Payment.Method;
 import com.maxmind.minfraud.request.Payment.Processor;
 import org.junit.jupiter.api.Test;
 
 public class PaymentTest {
+
+    @Test
+    public void testMethod() {
+        Payment payment = new Builder().method(Method.CARD).build();
+        assertEquals(Method.CARD, payment.getMethod());
+
+        payment = new Builder().method(Method.DIGITAL_WALLET).build();
+        assertEquals(Method.DIGITAL_WALLET, payment.getMethod());
+
+        payment = new Builder().method(Method.BUY_NOW_PAY_LATER).build();
+        assertEquals(Method.BUY_NOW_PAY_LATER, payment.getMethod());
+    }
 
     @Test
     public void testProcessor() {